+++ dialect/mysql/view.go
@@ -14,21 +14,21 @@ import (
 )
 
 func NewView[T any, C bob.Expression](tableName string, columns C) *View[T, []T, C] {
-	return NewViewx[T, []T](tableName, columns, nil)
-}
-
-// NewViewx creates a new View with a custom scanner.
-// If scanner is nil, it falls back to [scan.StructMapper].
-func NewViewx[T any, Tslice ~[]T, C bob.Expression](tableName string, columns C, scanner scan.Mapper[T]) *View[T, Tslice, C] {
-	v, _ := newView[T, Tslice](tableName, columns, scanner)
+	v, _ := newViewWithMapper[T, []T](tableName, columns, scan.StructMapper[T]())
 	return v
 }
 
-func newView[T any, Tslice ~[]T, C bob.Expression](tableName string, columns C, scanner scan.Mapper[T]) (*View[T, Tslice, C], mappings.Mapping) {
+// NewViewx creates a new View with a required custom scanner.
+func NewViewx[T any, Tslice ~[]T, C bob.Expression](tableName string, columns C, scanner scan.Mapper[T]) *View[T, Tslice, C] {
 	if scanner == nil {
-		scanner = scan.StructMapper[T]()
+		panic("mysql: nil mapper passed to NewViewx")
 	}
 
+	v, _ := newViewWithMapper[T, Tslice](tableName, columns, scanner)
+	return v
+}
+
+func newViewWithMapper[T any, Tslice ~[]T, C bob.Expression](tableName string, columns C, scanner scan.Mapper[T]) (*View[T, Tslice, C], mappings.Mapping) {
 	mappings := mappings.GetMappings(reflect.TypeOf(*new(T)))
 
 	return &View[T, Tslice, C]{
@@ -85,12 +85,10 @@ func (v *View[T, Tslice, C]) ColumnsExpr() expr.ColumnsExpr {
 // Query starts a select query on the view
 func (v *View[T, Tslice, C]) Query(queryMods ...bob.Mod[*dialect.SelectQuery]) *ViewQuery[T, Tslice] {
 	q := &ViewQuery[T, Tslice]{
-		Query: orm.Query[*dialect.SelectQuery, T, Tslice, bob.SliceTransformer[T, Tslice]]{
-			ExecQuery: orm.ExecQuery[*dialect.SelectQuery]{
-				BaseQuery: Select(sm.From(v.NameAsExpr())),
-				Hooks:     &v.SelectQueryHooks,
-			},
-			Scanner: v.scanner,
+		ModelQuery: orm.ModelQuery[*dialect.SelectQuery, T, Tslice]{
+			BaseQuery: Select(sm.From(v.NameAsExpr())),
+			Hooks:     &v.SelectQueryHooks,
+			Scanner:   v.scanner,
 		},
 	}
 
@@ -109,7 +107,7 @@ func (v *View[T, Tslice, C]) Query(queryMods ...bob.Mod[*dialect.SelectQuery]) *
 }
 
 type ViewQuery[T any, Ts ~[]T] struct {
-	orm.Query[*dialect.SelectQuery, T, Ts, bob.SliceTransformer[T, Ts]]
+	orm.ModelQuery[*dialect.SelectQuery, T, Ts]
 }
 
 func (v *ViewQuery[T, Ts]) Clone() *ViewQuery[T, Ts] {
@@ -117,9 +115,7 @@ func (v *ViewQuery[T, Ts]) Clone() *ViewQuery[T, Ts] {
 		return nil
 	}
 
-	return &ViewQuery[T, Ts]{
-		Query: v.Query.Clone(),
-	}
+	return &ViewQuery[T, Ts]{ModelQuery: v.ModelQuery.Clone()}
 }
 
 func (v *ViewQuery[T, Ts]) With(queryMods ...bob.Mod[*dialect.SelectQuery]) *ViewQuery[T, Ts] {