+++ dialect/sqlite/table.go
@@ -17,20 +17,27 @@ import (
 
 type (
 	setter[T any]                      = orm.Setter[T, *dialect.InsertQuery, *dialect.UpdateQuery]
-	ormInsertQuery[T any, Tslice ~[]T] = orm.Query[*dialect.InsertQuery, T, Tslice, bob.SliceTransformer[T, Tslice]]
-	ormUpdateQuery[T any, Tslice ~[]T] = orm.Query[*dialect.UpdateQuery, T, Tslice, bob.SliceTransformer[T, Tslice]]
-	ormDeleteQuery[T any, Tslice ~[]T] = orm.Query[*dialect.DeleteQuery, T, Tslice, bob.SliceTransformer[T, Tslice]]
+	ormInsertQuery[T any, Tslice ~[]T] = orm.ModelQuery[*dialect.InsertQuery, T, Tslice]
+	ormUpdateQuery[T any, Tslice ~[]T] = orm.ModelQuery[*dialect.UpdateQuery, T, Tslice]
+	ormDeleteQuery[T any, Tslice ~[]T] = orm.ModelQuery[*dialect.DeleteQuery, T, Tslice]
 )
 
 func NewTable[T any, Tset setter[T], C bob.Expression](schema, tableName string, columns C) *Table[T, []T, Tset, C] {
-	return NewTablex[T, []T, Tset](schema, tableName, columns, nil)
+	return newTableWithMapper[T, []T, Tset](schema, tableName, columns, scan.StructMapper[T]())
 }
 
-// NewTablex creates a new Table with a custom scanner.
-// If scanner is nil, it falls back to [scan.StructMapper].
+// NewTablex creates a new Table with a required custom scanner.
 func NewTablex[T any, Tslice ~[]T, Tset setter[T], C bob.Expression](schema, tableName string, columns C, scanner scan.Mapper[T]) *Table[T, Tslice, Tset, C] {
+	if scanner == nil {
+		panic("sqlite: nil mapper passed to NewTablex")
+	}
+
+	return newTableWithMapper[T, Tslice, Tset](schema, tableName, columns, scanner)
+}
+
+func newTableWithMapper[T any, Tslice ~[]T, Tset setter[T], C bob.Expression](schema, tableName string, columns C, scanner scan.Mapper[T]) *Table[T, Tslice, Tset, C] {
 	setMapping := mappings.GetMappings(reflect.TypeOf(*new(Tset)))
-	view, mappings := newView[T, Tslice](schema, tableName, columns, scanner)
+	view, mappings := newViewWithMapper[T, Tslice](schema, tableName, columns, scanner)
 	t := &Table[T, Tslice, Tset, C]{
 		View:          view,
 		pkCols:        expr.NewColumnsExpr(mappings.PKs...).WithParent(view.alias),
@@ -69,11 +76,9 @@ func (t *Table[T, Tslice, Tset, C]) PrimaryKey() expr.ColumnsExpr {
 // Starts an insert query for this table
 func (t *Table[T, Tslice, Tset, C]) Insert(queryMods ...bob.Mod[*dialect.InsertQuery]) *ormInsertQuery[T, Tslice] {
 	q := &ormInsertQuery[T, Tslice]{
-		ExecQuery: orm.ExecQuery[*dialect.InsertQuery]{
-			BaseQuery: Insert(im.Into(t.NameAsExpr())),
-			Hooks:     &t.InsertQueryHooks,
-		},
-		Scanner: t.scanner,
+		BaseQuery: Insert(im.Into(t.NameAsExpr())),
+		Hooks:     &t.InsertQueryHooks,
+		Scanner:   t.scanner,
 	}
 
 	q.Expression.AppendContextualModFunc(
@@ -93,11 +98,9 @@ func (t *Table[T, Tslice, Tset, C]) Insert(queryMods ...bob.Mod[*dialect.InsertQ
 // Starts an Update query for this table
 func (t *Table[T, Tslice, Tset, C]) Update(queryMods ...bob.Mod[*dialect.UpdateQuery]) *ormUpdateQuery[T, Tslice] {
 	q := &ormUpdateQuery[T, Tslice]{
-		ExecQuery: orm.ExecQuery[*dialect.UpdateQuery]{
-			BaseQuery: Update(um.Table(t.NameAsExpr())),
-			Hooks:     &t.UpdateQueryHooks,
-		},
-		Scanner: t.scanner,
+		BaseQuery: Update(um.Table(t.NameAsExpr())),
+		Hooks:     &t.UpdateQueryHooks,
+		Scanner:   t.scanner,
 	}
 
 	q.Expression.AppendContextualModFunc(
@@ -117,11 +120,9 @@ func (t *Table[T, Tslice, Tset, C]) Update(queryMods ...bob.Mod[*dialect.UpdateQ
 // Starts a Delete query for this table
 func (t *Table[T, Tslice, Tset, C]) Delete(queryMods ...bob.Mod[*dialect.DeleteQuery]) *ormDeleteQuery[T, Tslice] {
 	q := &ormDeleteQuery[T, Tslice]{
-		ExecQuery: orm.ExecQuery[*dialect.DeleteQuery]{
-			BaseQuery: Delete(dm.From(t.NameAsExpr())),
-			Hooks:     &t.DeleteQueryHooks,
-		},
-		Scanner: t.scanner,
+		BaseQuery: Delete(dm.From(t.NameAsExpr())),
+		Hooks:     &t.DeleteQueryHooks,
+		Scanner:   t.scanner,
 	}
 
 	q.Expression.AppendContextualModFunc(