+++ dialect/sqlite/wm/qm.go
@@ -59,7 +59,8 @@ func FromPreceding(exp any) bob.Mod[*clause.Window] {
 		w.SetStart(bob.ExpressionFunc(
 			func(ctx context.Context, w io.StringWriter, d bob.Dialect, start int) ([]any, error) {
 				return bob.ExpressIf(ctx, w, d, start, exp, true, "", " PRECEDING")
-			}),
+			},
+		),
 		)
 	})
 }
@@ -75,7 +76,8 @@ func FromFollowing(exp any) bob.Mod[*clause.Window] {
 		w.SetStart(bob.ExpressionFunc(
 			func(ctx context.Context, w io.StringWriter, d bob.Dialect, start int) ([]any, error) {
 				return bob.ExpressIf(ctx, w, d, start, exp, true, "", " FOLLOWING")
-			}),
+			},
+		),
 		)
 	})
 }
@@ -85,7 +87,8 @@ func ToPreceding(exp any) bob.Mod[*clause.Window] {
 		w.SetEnd(bob.ExpressionFunc(
 			func(ctx context.Context, w io.StringWriter, d bob.Dialect, start int) ([]any, error) {
 				return bob.ExpressIf(ctx, w, d, start, exp, true, "", " PRECEDING")
-			}),
+			},
+		),
 		)
 	})
 }
@@ -101,7 +104,8 @@ func ToFollowing(exp any) bob.Mod[*clause.Window] {
 		w.SetEnd(bob.ExpressionFunc(
 			func(ctx context.Context, w io.StringWriter, d bob.Dialect, start int) ([]any, error) {
 				return bob.ExpressIf(ctx, w, d, start, exp, true, "", " FOLLOWING")
-			}),
+			},
+		),
 		)
 	})
 }