+++ gen/bobgen-sqlite/driver/parser/visitor.go
@@ -1007,7 +1007,8 @@ func (v *visitor) VisitExpr_bind(ctx *sqliteparser.Expr_bindContext) any {
 				EditedPosition: [2]int{start, end},
 			})
 			return nil
-		}),
+		},
+	),
 	)
 
 	return nil
@@ -1233,7 +1234,8 @@ func (v *visitor) VisitSelect_stmt(ctx *sqliteparser.Select_stmtContext) any {
 		if len(source.Columns) != len(coreSource.Columns) {
 			v.Err = fmt.Errorf(
 				"select core %d: column count mismatch %d != %d",
-				i, len(source.Columns), len(coreSource.Columns))
+				i, len(source.Columns), len(coreSource.Columns),
+			)
 			return nil
 		}
 
@@ -1243,7 +1245,8 @@ func (v *visitor) VisitSelect_stmt(ctx *sqliteparser.Select_stmtContext) any {
 			if len(source.Columns[i].Type) == 0 {
 				v.Err = fmt.Errorf(
 					"select core %d: column %d type mismatch:\n%v\n%v",
-					i, i, col.Type, coreSource.Columns[i].Type)
+					i, i, col.Type, coreSource.Columns[i].Type,
+				)
 				return nil
 			}